… addons/account/tests/test_account_move_entry.py
… addons/account/tests/test_invoice_taxes.py
… addons/account/views/account_move_views.xml