[WIP] Use computed fields instead of onchange, no more ORM hacks: Proof of Concept

Performance comparison: (~2x faster at creation & line onchange)

1/ create invoice with 2000 lines, 1 tax per line:
- master:        37s
- this branch: 17s

2/ onchange invoice line (measured on 10th line):
- master:        228ms (5 rpc queries)
- this branch: 106ms (4 RPC queries)




--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr